Added in new viewmodels. Dependency injention
Added in the following viewmodels:

BaseViewModel: Base class for all viewmodels to inherit from, includes observable properties that can and will be used in multiple view models.

LoginViewModel: Viewmodel for the login page, manages the logic of navigating to the register page and logging in.

RegisterViewModel: Handles the logic of registering a new user and navigating back to the login page.

Added in the nugget package CommunityToolkit.Mvvm by Microsoft.

This package is used to reduce the amount of boilerplate code to achieve synchronous UI changes though the help of different observable properties.

The BaseViewModel inherits from ObservableObject which uses the package. All other viewmodels will also get the same benefits through inheriting from the BaseViewModel.