Some UI changes to RegisterPage, Mediapicker
Some adjustments to the UI of the RegisterPage.

Added a new API called Mediapicker and dependency injected the RegisterViewModel with the new API.

Mediapicker is used to let the users be able to select images.

The user can now select a image of their liking that will be used as a profile picture, the selected image is at the moment only displayed in the RegisterPage with no additional functionality.