Bug workaround until Microsoft fixes it
There is a bug where if your page is of the type singleton instead of transient and you have frames inside the page, the app crashes when navigating in and out of the page.

This is a known bug that has not been fixed yet but there are two ways of resolving this bug.

First way is to let the page be of the type transient. Second way is to not use frames but instead borders in your Xaml file.

Added a comment in the RegisterPage.Xaml that describes this bug.