DatabaseManager, SQLite
Created a new class called DatabaseManager in the DataAccess folder. This class initializes the database locally and any given tables that are specified, if the database/tables do not already exist. This class is also used to perform CRUD operations on the database. This is done with the SQLite database system framework.

Installed a package called SQLite-net-pcl that is used to be able to create and use a database, regardless of which platform the application is used on. Its also one of the most popular relational database systems out there.

Created a new model called Test, used for testing the SQLite database.

Created 2 tables in the database and one insert operation, this was done primarily for testing purposes.