Registering Users, RegisterService, Dependency Injection, Databinding
Users can now register their accounts in the RegisterPage. This is not yet completely finished as hashing, creating unique salts, and checking that there is not another user that has registered with  the same mail will be implemented.

Created a new service called RegisterService and dependency injected it into the RegisterViewModel. The purpose of the service is to manage the registration of new users.

Added new fields: string username, password, name, lastName, and email in the RegisterViewModel. And databinded these fields to the entrys in RegisterPage.xaml

Also removed the TEST model and dropped the related table from the database.