Activity indicator, Styles.xaml
Added an activity indicator that is enabled based on IsBusy,.

The styles Button_General and Entry_General in Styles.xaml have a new setter property on IsEnabled that is based on the BaseViewModel IsNotEnabled.

bool IsNotBusy is the reverse of bool IsBusy, IsBusy gets set to true when different actions are made, like for example registering a new user to the database.