Hash and Salt, Logging in , And more..
Created a new service called SecurityService, this service manages the hashing of passwords, generation of salts , and verification of passwords.

Created a new service called LoginService, at the moment it only calls to the DatabaseManager.cs to return a user from a specified username.

Created a new method called GetUserByUsername inside of DatabaseManager.cs that querys to the database table of User and returns a User if the username specified exist in the table.

The hashing algorithm chosen was Argon2 as its allegedly one of the best if not the best hashing algorithm out there at the moment, installed a new package called Konscious.Security.Cryptography.Argon2  that is used to be able to do the argon2 hashing algorithm.

A new string field called Salt is now stored in the table of User in the database.

When a user now registers a new password a random salt with the size of 16 (size can be changed) is generated and stored to the user. And the argon2 hashing algorithm is applied to the entered password.

Dependency injected LoginService to the LoginViewModel and injected the SecurityService to both the LoginViewModel and RegisterViewModel.

Also registered the new services in the MauiProgram.cs file.

Added a new style of ActivityIndicator called ActivityIndicator _General in Styles.xaml.

Some small changes to the UI in LoginPage.