Increase of argon2.DegreeOfParallelism
Increased the argon2.DegreeOfParallelism to 16, this makes the mobile/desctop that its deployed on to use more CPU cores while the hashing algorithm is in use.

Tested the registration and login on my physical android phone and it takes about 3-4 seconds for both actions to complete, which is pretty good as argon2 is a very expensive algorithm.