Username recovery,RecoveryViewModel, EmailService, Dependency Injection
Created a new viewmodel called RecoveryViewModel and dependency injected a new service called EmailService into the view model and databinded this to a new page called RecoveryUsernamePage.

Purpose of  RecoveryViewmodel is to gather the email that the user inputs in the RecoverUserNamePage and then use the dependency injected EmailService to send an email , the email only gets sent if it exist in the database.

When the user has forgotten their username and want to recover it they can click on a link label in the LoginPage that then navigates the user to the RecoverUserNamgePage. There the user inputs their email address and gets the username details sent to their email.

EmailService uses SmtpClient and  MailMessage as the logic to be able to send emails to users, the emails sent come from my project email outlook account.