Refactoring, DatabaseUserService,
Did some refactoring in the form of removing the DataAccess folder with the public static class of DatabaseManager.cs and made a new service called DatabaseUserService with the same code but while not being static. Reasonings behind this is that in MVVMS, a common approach is to use services to encapsulate data access and business logic. It can make the code more modular and easier to maintain. If the project is expected to grow and you anticipate the need for more complex data access or if you want to follow best practices for testability, its best to use a service.

Depency injected the DatabaseUserService into relevant viewmodels and refactored the code where it was needed as the old static class is gone.

Registred the DatabaseUserService as singleton in the MauiProgram.cs file, as its a service that is and will be used often, removing the need to create a new instance of the service every time its referenced in a viewmodel.

Removed LoginService as I dont find a need for it at the moment, will create a new one I find a purpose of having a LoginService. The DatabaseUserService and SecurityService takes care of everything needed to login a  user.

Renamed RecoveryViewModel to RecoverUsernameViewModel.

Created a new viewmodel called RecoverPasswordVIewModel and dependency injected these services into the viewmodel ; SecurityService and EmailService. This viewmodel with its corelated services will manage the process of recovering a users password, its not completed yet.