Reset user password,
Changed the name of RecoverPasswordViewModel to ResetPasswordViewmodel as its a more accurate name for the purpose.

Created three new pages: ResetPasswordEmailPage, ResetPasswordTokenPage, and ResetPasswordPage. Registered these in the MauiProgram.cs file and binded them to the ResetPasswordViewModel.

Added two new fields to the User model : string PasswordResetToken and DateTime PasswordResetDate.

The user can now click on the link label "Forgot Password" in LoginPage to get navigated to ResetPasswordEmailPage.

ResetPasswordViewModel alongside DatabaseUserService, SecurityService, and EmailService now handle the procedure of resetting the user password. The user enters their email address in the ResetPasswordEmailPage, in the viewmodel the email is verified if it exist in the sqlite database, if it exist then a hashed password reset token gets assigned to the user as well as the current date. Then an email gets sent to the user with the token.

After that the user gets navigated to the ResetPasswordTokenPage. There the user has to enter the token that was sent to their email address, if the token the user types in matches the hashed version of the token in the database and if five minutes has not passed since the start of the reset, the user gets navigated to the ResetPasswordPage.

Here the user gets to create their new password as long as the token is still valid. After the user has created their new password the database then updates the changes made and the user gets returned to the LoginPage.