User Profile Picture, Image Compression, Image Storing
Added a new public byte[] ProfilePicture { get; set; } field to the User model.

Added a new class into the Utility folder called ImageTool.

Installed a new nugget package called SkiaSharp, this package is used to be able to optimize images through compression.

Created two new methods inside the ImageTool class called CompressAndResizeImage and CalculateTargetDimensions. The CompressAndResizeImage  handles the compression of a specified byte[] imageStream to a specifik width,height, and quality, this is done with the help of the SkiaSharp package. The CalculateTargetDimensions handles the correct target dimensions depending on what the width/height parameters of the CompressAndResizeImage have been set to.

Inside the RegisterViewModel a new private byte[] profilePicture { get; set; } has been created. This variable is set to the image the user picks in the RegisterPage. When the user then tries to register the variable gets optimized through the help of the ImageTool class and then gets set to the User which then finally gets added to the database.