AppShell changes, UI Change after login
Added a new route to a TabBar in the AppShell.xaml. Added som temporary tab icons and pages for the tabs.

Added some new colors in Colors.xaml

When the user now logs in, the AppShell changes route to the TabBar route which is called "Home", here the user will be able to easily navigate to the main pages with the help of tabs.