Images on physical phones <-> SkiaSharp, Admin login
While using a physical device and selecting a profile image on the register page and viewing it, the image got weirdly rotated. The issue was that the images on physical phone devices contained specific OS metadata that affected the rotation of said images. Solved this issue by creating a SKImage from the file fullpath and then creating a SKBitmap from the SKImage, doing this resolved the rotation issue as SkImage makes right corrections based on the metadata from the image.

Made a new instant login through using the username and password of "admin". The only reason for doing this is because of the hashing algorithm taking a long time on my android emulators. Take note that the slow speed only applies for the emulator, as using hashing on physical devices does not result in this issue. Also tested the BCrypt hashing algoritm as its quite a lightweight hashing algorithm but the issue still persisted on the emulator.