UserProfilePage, UserProfileViewMode,, UserService, And more...
Created a new page called UserProfilePage, where the user can see their profile image, username, name, last name, and email. The plan is that the user will be able to change these attributes to their liking. Currently the user can only change their profile picture.

Created a new UserProfileViewModel, bound it to the UserProfilePage, and dependency injected DatabaseUserService, UserService, and IMediaPicker. This viewmodels purpose is to manage how the user will be able to change their attributes by interacting with the UserProfilePage.

Created a new service called UserService. The service holds the logged in user details and is of the singleton type, which means that it is only created once and only then referenced when it is needed. The purpose of this service is to hold the logged in user details and to let other viewmodels get the necessary user details through dependency injection.