Change passsword/email, Refactoring and fixes, New pages, With more...
The logged in user can now change their password and email in the UserProfilePage.

Created new pages called ChangePasswordPage, ChangePasswordSecurityCheckPage, ChangeEmailCurrentEmailTokenPage, ChangeEmailNewEmailPage, ChangeEmailNewEmailTokenPage, and finished the UI for them. Added new routings to the pages and registered the pages in AppShell.xaml.cs respective MauiProgram.cs. Bound the UserProfileViewModel to the pages.

I had forgotten to update the database when the user reset their password in the ResetPasswordPage, this has been fixed now.

When the user changes their password an email gets sent to their email address, notifying the user that their password has been changed. The same principle for when the user changes their email.

Some refactoring.