Logout ,AppShell changes, Tweaks
The user can now on the UserProfilePage logout, this resets the UserService.user and returns them to the LoginPage.

Seperated the old AppShell into two new ones called AppShellLogin and AppShellMain. the AppSHellLogin contains only a shellcontent for the loginpage, and AppShellMain contains a tabbar. Modified the routings for each AppShell.

Initially the appShell will be AppShellLogin and when the user logs in successfully the appshell changes to AppShellMain and vice versa. This ensures that the code is modular as you have basically two different UI structures of the application and it also makes sure that the transient data in the tabbar of AppShellMain gets called again as in having both a tabbar and a shellontent within an appshell (and navigating between them using routes) results in the tabbar saving the data of the pages within the tabbar regardless of the page/viewmodel is transient or singleton.

Made small changes/tweaks to the UI of some pages.