ProductCreationPage developments, ProductFactoryService, ProductCategoryEnum, ProductItemEnum
Added two new pickers to the ProductCreationPage where the user can select the category of the product and the product item that they are planning to put out in the market.

Added a new Service called ProductFactoryService. It manages the process of when the user picks/creates products.

Dependency injected ProductFactoryService into the ProductCreationViewModel.

Added two new enum classes called ProductCategoryEnum and ProductItemEnum.