ChangeEmailViewModel, Messenger,Refactoring
Created a new viewmodel called ChangeEmailViewModel, this viewmodel handles the process of the user changing their email adress.

Swapped this out from the old system that was used in UserProfileViewModel as I thought that it is ineffiecient to have multiple pages call on UserProfileViewModel to be able to change an email adress.

I now instead use the new ChangeEmailViewModel with only a single page do do the same thing. Which now results in more modularity and improved performance/memory.

Removed the old methods from UserProfileViewModel.

Removed the old pages that were related to changing the email adress with all relevant bindings/routings...

Also refactored some parts of changing the email adress.

Created two new message classes, one called UpdateUserItemMessage and the other ReceiveUserItemMessage, the latter is not currently being used.

I now use .NET MAUI messaging system (pub/sub) to notify from ChangeEmailViewModel that the email has been changed which then calls a method in UserProfileViewModel to call OnPropertyChanged for the user, so that the updated email can be seen when returning to the profile page. The UserProfileViewModel now subscribes to the UpdateUserItemMessage and the ChangeEmailViewModel sends the message, which instantly runs a method in UserProfileViewModel .

Some misc changes.