ChangePasswordViewModel, Additions & Refactoring
Made a new viewmodel called ChangePasswordViewModel which manages the process of an logged in user changing their password.

Removed the old method that were used to change the users password in UserProfileViewModel and moved them to the ChangePasswordViewModel,.

Removed ChangePasswordSecurityCheckPage.

Improved and did some refactorings to the methods regarding changing the users password.

Changed the ChangePasswordPage.cs view to now also have the contents of ChangePasswordSecurityCheckPage. Binded ChangePasswordViewModel to the page, the viewmodel handles the contents that are shown using databinding.

Misc Changes.